# examples/demo-backend/openapi/complex.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/notifications": {
      "get": {
        "operationId": "Notifications_list",
        "parameters": [
          {
            "name": "X-Demo-Scenario",
            "in": "header",
            "description": "Forces a deterministic frontend-testing scenario for this request.",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "normal",
                "slow",
                "timeout",
                "server-error",
                "rate-limited",
                "empty",
                "expired-auth",
                "forbidden",
                "conflict",
                "large-dataset"
              ]
            }
          },
          {
            "name": "X-Organization-Id",
            "in": "header",
            "description": "Current tenant. The authenticated user must be a member.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "org-acme"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "example": "notification-020",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationsResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is missing or expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "The current user lacks permission.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Demo rate limit scenario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected or simulated server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieSession": []
          }
        ],
        "summary": "List polymorphic notifications using cursor pagination",
        "tags": [
          "Notifications"
        ]
      }
    },
    "/api/v1/notifications/{id}/read": {
      "post": {
        "operationId": "Notifications_markRead",
        "parameters": [
          {
            "name": "X-Demo-Scenario",
            "in": "header",
            "description": "Forces a deterministic frontend-testing scenario for this request.",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "normal",
                "slow",
                "timeout",
                "server-error",
                "rate-limited",
                "empty",
                "expired-auth",
                "forbidden",
                "conflict",
                "large-dataset"
              ]
            }
          },
          {
            "name": "X-Organization-Id",
            "in": "header",
            "description": "Current tenant. The authenticated user must be a member.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "org-acme"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is missing or expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "The current user lacks permission.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Demo rate limit scenario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected or simulated server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieSession": []
          },
          {
            "csrf": []
          }
        ],
        "summary": "Mark a notification as read",
        "tags": [
          "Notifications"
        ]
      }
    },
    "/api/v1/files": {
      "get": {
        "operationId": "Files_list",
        "parameters": [
          {
            "name": "X-Demo-Scenario",
            "in": "header",
            "description": "Forces a deterministic frontend-testing scenario for this request.",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "normal",
                "slow",
                "timeout",
                "server-error",
                "rate-limited",
                "empty",
                "expired-auth",
                "forbidden",
                "conflict",
                "large-dataset"
              ]
            }
          },
          {
            "name": "X-Organization-Id",
            "in": "header",
            "description": "Current tenant. The authenticated user must be a member.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "org-acme"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilesResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is missing or expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "The current user lacks permission.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Demo rate limit scenario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected or simulated server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieSession": []
          }
        ],
        "summary": "List uploaded file metadata",
        "tags": [
          "Files"
        ]
      },
      "post": {
        "operationId": "Files_upload",
        "parameters": [
          {
            "name": "X-Demo-Scenario",
            "in": "header",
            "description": "Forces a deterministic frontend-testing scenario for this request.",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "normal",
                "slow",
                "timeout",
                "server-error",
                "rate-limited",
                "empty",
                "expired-auth",
                "forbidden",
                "conflict",
                "large-dataset"
              ]
            }
          },
          {
            "name": "X-Organization-Id",
            "in": "header",
            "description": "Current tenant. The authenticated user must be a member.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "org-acme"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is missing or expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "The current user lacks permission.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Demo rate limit scenario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected or simulated server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieSession": []
          },
          {
            "csrf": []
          }
        ],
        "summary": "Upload a file up to 5 MiB and retain it in memory",
        "tags": [
          "Files"
        ]
      }
    },
    "/api/v1/files/{id}/download": {
      "get": {
        "operationId": "Files_download",
        "parameters": [
          {
            "name": "X-Demo-Scenario",
            "in": "header",
            "description": "Forces a deterministic frontend-testing scenario for this request.",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "normal",
                "slow",
                "timeout",
                "server-error",
                "rate-limited",
                "empty",
                "expired-auth",
                "forbidden",
                "conflict",
                "large-dataset"
              ]
            }
          },
          {
            "name": "X-Organization-Id",
            "in": "header",
            "description": "Current tenant. The authenticated user must be a member.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "org-acme"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Binary file content.",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or validation error.",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is missing or expired.",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "The current user lacks permission.",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource does not exist.",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Demo rate limit scenario.",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected or simulated server error.",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieSession": []
          }
        ],
        "summary": "Download an in-memory file",
        "tags": [
          "Files"
        ]
      }
    },
    "/api/v1/audit-events": {
      "get": {
        "operationId": "Audit_list",
        "parameters": [
          {
            "name": "X-Demo-Scenario",
            "in": "header",
            "description": "Forces a deterministic frontend-testing scenario for this request.",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "normal",
                "slow",
                "timeout",
                "server-error",
                "rate-limited",
                "empty",
                "expired-auth",
                "forbidden",
                "conflict",
                "large-dataset"
              ]
            }
          },
          {
            "name": "X-Organization-Id",
            "in": "header",
            "description": "Current tenant. The authenticated user must be a member.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "org-acme"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "action",
            "required": false,
            "in": "query",
            "schema": {
              "example": "product.updated",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditEventsResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is missing or expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "The current user lacks permission.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Demo rate limit scenario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected or simulated server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieSession": []
          }
        ],
        "summary": "List immutable audit events using offset pagination",
        "tags": [
          "Audit"
        ]
      }
    },
    "/api/v1/exports/orders": {
      "post": {
        "operationId": "Jobs_startExport",
        "parameters": [
          {
            "name": "X-Demo-Scenario",
            "in": "header",
            "description": "Forces a deterministic frontend-testing scenario for this request.",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "normal",
                "slow",
                "timeout",
                "server-error",
                "rate-limited",
                "empty",
                "expired-auth",
                "forbidden",
                "conflict",
                "large-dataset"
              ]
            }
          },
          {
            "name": "X-Organization-Id",
            "in": "header",
            "description": "Current tenant. The authenticated user must be a member.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "org-acme"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is missing or expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "The current user lacks permission.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Demo rate limit scenario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected or simulated server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieSession": []
          },
          {
            "csrf": []
          }
        ],
        "summary": "Start an asynchronous orders export",
        "tags": [
          "Background jobs"
        ]
      }
    },
...
  },
  "info": {
    "title": "Demo Complex API",
    "description": "Cookie-session, multitenant and realtime API for large frontend applications.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3002",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "demo_session"
      },
      "csrf": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Required for authenticated mutations."
      }
    },
    "schemas": {
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 404
          },
          "code": {
            "type": "string",
            "example": "PRODUCT_NOT_FOUND"
          },
          "message": {
            "type": "string",
            "example": "Product not found"
          },
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErrorDetailDto"
            }
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2026-07-30T12:00:00.000Z"
          },
          "path": {
            "type": "string",
            "example": "/api/v1/products/product-404"
          },
          "requestId": {
            "type": "string",
            "example": "req-5c9f7a3d"
          }
        },
        "required": [
          "statusCode",
          "code",
          "message",
          "details",
          "timestamp",
          "path",
          "requestId"
        ]
      },
...
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "NotificationsResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NotificationDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/CursorMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "NotificationResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/NotificationDto"
          }
        },
        "required": [
          "data"
        ]
      },
...
      "FilesResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FileMetadataDto"
            }
          }
        },
        "required": [
          "data"
        ]
      },
      "FileResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/FileMetadataDto"
          }
        },
        "required": [
          "data"
        ]
      },
...
      "AuditEventsResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuditEventDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PageMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
      "JobResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/JobDto"
          }
        },
        "required": [
          "data"
        ]
      },
...
    }
  }
}